Add the `dish` and `sink` LWRPs.
A `dish` is a triple kept in attribute state that contains:
* A `sink` label
* A "dish" `recipe`
* A `wash_recipe`.

It does not represent anything on the node itself: it is represented completely in chef server state.

A dish may be clean or dirty.  It is clean if either its `recipe` or `wash_recipe` are present in the node's recipe list, and dirty if neither are present.

Dishes support two actions.  The default, `:use`, ensures that the triple is present in the chef server's memory.  The `:toss` action ensures that it is absent instead.

A `sink` is a collection of dishes.  Dishes choose their sink with the `sink` attribute, with default `default`.  Sinks have a single action, `:wash`.  It considers each contained dish and, if it is dirty, `include_recipe`s the dish's `wash_recipe`.